Move unit-testable diagnostics logic to core
